[chore] 로그인 버튼 속 사진 유저 상호작용 해제
